refactor: move order quotes logic
